test: restructure overlay unit tests for better readability (CP: 25.1)
