+++ packages/overlay/test/multiple.test.js
@@ -1,5 +1,5 @@
 import { expect } from '@vaadin/chai-plugins';
-import { click, escKeyDown, fixtureSync, nextRender } from '@vaadin/testing-helpers';
+import { fixtureSync, nextRender } from '@vaadin/testing-helpers';
 import sinon from 'sinon';
 import '../src/vaadin-overlay.js';
 import { createOverlay } from './helpers.js';
@@ -48,118 +48,6 @@ describe('multiple overlays', () => {
         expect(overlay1._last).to.be.true;
       });
     });
-
-    describe('vaadin-overlay-escape-press', () => {
-      let spy;
-
-      beforeEach(() => {
-        spy = sinon.spy();
-        overlay1.addEventListener('vaadin-overlay-escape-press', spy);
-      });
-
-      it('should fire the vaadin-overlay-escape-press if it is the only overlay opened', () => {
-        overlay1.opened = true;
-        escKeyDown(document.body);
-        expect(spy.called).to.be.true;
-      });
-
-      it('should not fire the vaadin-overlay-escape-press if there is a recent overlay opened', () => {
-        overlay1.opened = true;
-
-        overlay2.opened = true;
-
-        escKeyDown(document.body);
-        expect(spy.called).to.be.false;
-      });
-    });
-
-    describe('vaadin-overlay-outside-click', () => {
-      let spy;
-
-      beforeEach(() => {
-        spy = sinon.spy();
-        overlay1.addEventListener('vaadin-overlay-outside-click', spy);
-      });
-
-      it('should fire the vaadin-overlay-outside-click if it is the only overlay opened', () => {
-        overlay1.opened = true;
-        click(parent);
-        expect(spy.calledOnce).to.be.true;
-      });
-
-      it('should not fire the vaadin-overlay-outside-click if there is a recent overlay opened', () => {
-        overlay1.opened = true;
-
-        overlay2.opened = true;
-
-        click(parent);
-        expect(spy.called).to.be.false;
-      });
-    });
-
-    describe('pointer-events', () => {
-      it('should restore pointer-events correctly when overlays are not closed in order', () => {
-        overlay1.opened = true;
-
-        overlay2.opened = true;
-        expect(document.body.style.pointerEvents).to.eql('none');
-
-        overlay1.opened = false;
-
-        overlay2.opened = false;
-        expect(document.body.style.pointerEvents).to.eql('');
-      });
-
-      it('should disable pointer-events in first overlay when second opens', () => {
-        const overlay1Part = overlay1.shadowRoot.querySelector('[part="overlay"]');
-        overlay1.opened = true;
-
-        overlay2.opened = true;
-
-        expect(getComputedStyle(overlay1Part).pointerEvents).to.equal('none');
-      });
-
-      it('should restore pointer-events in first overlay when second closes', () => {
-        const overlay1Part = overlay1.shadowRoot.querySelector('[part="overlay"]');
-        overlay1.opened = true;
-
-        overlay2.opened = true;
-
-        overlay2.opened = false;
-        expect(getComputedStyle(overlay1Part).pointerEvents).to.equal('auto');
-      });
-
-      it('should restore pointer-events in second overlay when third closes', () => {
-        const overlay1Part = overlay1.shadowRoot.querySelector('[part="overlay"]');
-        const overlay2Part = overlay2.shadowRoot.querySelector('[part="overlay"]');
-        overlay1.opened = true;
-
-        overlay2.opened = true;
-
-        overlay3.opened = true;
-
-        overlay3.opened = false;
-
-        expect(getComputedStyle(overlay2Part).pointerEvents).to.equal('auto');
-        expect(getComputedStyle(overlay1Part).pointerEvents).to.equal('none');
-      });
-
-      it('should clear pointer events after closing overlays', () => {
-        const overlay1Part = overlay1.shadowRoot.querySelector('[part="overlay"]');
-        // Step 1: Opening overlay 1 so it's physically moved under the body
-        overlay1.opened = true;
-        // Step 2: As overlay2 is modal, it will set overlay 1's pointer events to none
-        overlay2.opened = true;
-        // Step 3: Closing overlay 1 so it's physically moved back from under the body
-        overlay1.opened = false;
-        // Step 4: Closing overlay 2 restores pointer-events of an overlay it
-        // finds under the body node, but overlay 1 is no longer there.
-        overlay2.opened = false;
-        // The fix: Clear pointer-events whenever an overlay is closed
-        // (in this case overlay 1 at step 3)
-        expect(getComputedStyle(overlay1Part).pointerEvents).to.equal('auto');
-      });
-    });
   });
 
   describe('modeless', () => {
@@ -174,28 +62,12 @@ describe('multiple overlays', () => {
     };
 
     beforeEach(async () => {
-      parent = fixtureSync(`
-        <div id="parent">
-          <vaadin-overlay modeless></vaadin-overlay>
-          <vaadin-overlay modeless></vaadin-overlay>
-        </div>
-      `);
-      modeless1 = parent.children[0];
-      modeless1.renderer = (root) => {
-        if (!root.firstChild) {
-          root.textContent = 'overlay 1';
-          const input = document.createElement('input');
-          root.appendChild(input);
-        }
-      };
-      modeless2 = parent.children[1];
-      modeless2.renderer = (root) => {
-        if (!root.firstChild) {
-          root.textContent = 'overlay 2';
-          const input = document.createElement('input');
-          root.appendChild(input);
-        }
-      };
+      parent = fixtureSync('<div></div>');
+      modeless1 = createOverlay('overlay 1');
+      modeless2 = createOverlay('overlay 2');
+      modeless1.modeless = true;
+      modeless2.modeless = true;
+      parent.append(modeless1, modeless2);
       await nextRender();
     });
 
@@ -267,59 +139,6 @@ describe('multiple overlays', () => {
       expect(showSpy2).to.be.not.called;
 
       expect(modeless2._last).to.be.true;
-    });
-
-    it('should not fire the vaadin-overlay-escape-press if the overlay does not contain focus', () => {
-      const spy = sinon.spy();
-      modeless1.addEventListener('vaadin-overlay-escape-press', spy);
-
-      modeless1.opened = true;
-
-      escKeyDown(document.body);
-      expect(spy.called).to.be.false;
-    });
-
-    it('should not fire the vaadin-overlay-escape-press if the overlay contains focus', () => {
-      const spy = sinon.spy();
-      modeless1.addEventListener('vaadin-overlay-escape-press', spy);
-
-      modeless1.opened = true;
-
-      const input = modeless1.querySelector('input');
-      input.focus();
-
-      escKeyDown(input);
-      expect(spy.called).to.be.true;
-    });
-
-    it('should fire the vaadin-overlay-escape-press if the overlay is the frontmost one', () => {
-      const spy = sinon.spy();
-      modeless1.addEventListener('vaadin-overlay-escape-press', spy);
-
-      modeless1.opened = true;
-
-      modeless2.opened = true;
-      modeless1.bringToFront();
-
-      const input = modeless1.querySelector('input');
-      input.focus();
-
-      escKeyDown(input);
-      expect(spy.called).to.be.true;
-    });
-
-    it('should not fire the vaadin-overlay-escape-press if the overlay is not the frontmost', () => {
-      const spy = sinon.spy();
-      modeless1.addEventListener('vaadin-overlay-escape-press', spy);
-
-      modeless1.opened = true;
-      modeless2.opened = true;
-
-      const input = modeless2.querySelector('input');
-      input.focus();
-
-      escKeyDown(input);
-      expect(spy.called).to.be.false;
     });
   });
 });
